+++ CHANGELOG.md
@@ -1,2 +1,7 @@
-- Fixes a bug with RTDB Rules hot reloading in the RTDB emulator. (#2371)
-- Changes default functions runtime to Node.js 10 for `firebase init`.
+- Fixes issue where `Authorization` header was missing from callable functions in the emulator (#2459).
+- Improve support for the Node.js 12 (Beta) runtime in the Functions emulator.
+- Allow specifying the config (`firebase.json`) file using the `--config`/`-c` flag.
+- Fixes issue where `emulators:exec` could fail to shut down cleanly (#2477).
+- Fixes issue where database emulator did not properly load initial rules (#2483).
+- Allow starting the UI with `emulators:exec` using the `--ui` flag.
+- Fixes issue where multiple CLI instances compete for the same log (#2464).